Dont interfer in the removal of deliver-id mappings of sent messages
This fixes an issue where incoming messages could lock up a send request
because dropping a received Delivery removes the mapping for the Sender.
The SessionDriver will then not be able to notifiy the LinkDriver in
SessionDriver::dispatch => Performative::Disposition (line 438). The
Sender then might wait forever in Sender::send_withMode =>
self.link.recv() (line 637).

On another note: This seems also seems to be an error besides the
potential lockup of the Sender. Why does a received Delivery/Message
interfere in the mappings of sent messages at all?